requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
acvl_utils==0.2.5
antspyx==0.6.1
batchgenerators==0.25.1
batchgeneratorsv2==0.3.0
blosc2==3.6.1
brainextractor==0.2.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
connected-components-3d==3.24.0
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
dynamic_network_architectures==0.4.2
einops==0.8.1
fft-conv-pytorch==1.2.0
filelock==3.18.0
fonttools==4.59.0
freetype-py==2.5.1
fsspec==2025.7.0
future==1.0.0
graphviz==0.21
huggingface-hub==0.34.3
humanize==4.12.3
idna==3.10
imagecodecs==2025.3.30
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
linecache2==1.0.0
llvmlite==0.44.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
ndindex==1.10.0
networkx==3.4.2
nibabel==5.3.2
nilearn==0.12.0
nnunetv2==2.6.2
numba==0.61.2
numexpr==2.11.0
numpy==2.2.6
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
py-cpuinfo==9.0.0
pyglet==2.1.6
Pygments==2.19.2
PyOpenGL==3.1.0
pyparsing==3.2.3
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
simpleitk==2.5.2
six==1.17.0
statsmodels==0.14.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.19
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchio==0.20.19
torchvision==0.22.1
tqdm==4.67.1
traceback2==1.4.0
trimesh==4.7.1
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
unittest2==1.1.0
urllib3==2.5.0
webcolors==24.11.1
wrapt==1.17.2
yacs==0.1.8